Only throw Error if destination node specified
In single_source_shortest_paths, an Error would be thrown any time
a path to a destination node, `d`, wasn't found, but `d` is
optional--when it's not passed, paths from the start node to all
reachable nodes should be returned.

This is another bug found due to looking into issue #1.