Remove naive break-early logic
In single_source_shortest_paths(),` as soon as the destination node `d`
was visited, the algorithm would break early for performance reasons,
the logic being that there's no need to explore the rest of the graph if
`d` has already been found.

The problem with this approach is that it's possible to reach `d`
without having actually found the shortest path to it (duh!).

An approach that might work would be to record which nodes `d` has been
visited from and break early if `d` has been visited from *all* of it's
neighbors.

refs #1